new src/2d/dig-bouss for merging geoclaw/src/2d/bouss with dclcaw/src/2d/dig
First pass at `Makefile.dig-bouss` but the list of BOUSSLIB routines includes
repetitions and needs to be fixed after cleaning up `geoclaw/src/2d/bouss`.

The list of DIGBOUSS sources in the Makefile are the ones that were changed
in both dig and bouss, but maybe some of these can be cleaned up for more
general use (e.g. `valout.f90`).